backend/requirements.txt
# Core Backend Framework
fastapi
uvicorn[standard]

# ORM & Database
sqlalchemy
psycopg2-binary
pydantic
python-dotenv

# HTTP & API Clients
httpx
requests

# Google APIs
google-auth
google-auth-oauthlib
google-api-python-client

# Optional: Email/Notification Providers
slack_sdk
firebase-admin
sendgrid

# LLM Integration
openai  # if you want OpenAI fallback
# google-generativeai  # if using Gemini directly

# Bonus: Security & Auth
passlib[bcrypt]
python-jose[cryptography]

# Testing
pytest